Create separate resource for canvas
By moving the canvas context and element into a resource, we can
create multiple renderers that can all update the same canvas. Right
now it's implemented by wrapping the CanvasRenderingContext2d API and
using RefCells internally to avoid having to mutably borrow the
resource from the world, but this is probably not the best approach.